Remove AssumeMixin.
This allows us to get rid of some tricky magic in
ManagedProperties.
